Implemented approve and started making test cases

+++ Project 1/src/Possession.java
@@ -2,16 +2,16 @@
 public class Possession extends InanimateAccount
 {
 	private Person owner;
-	private double price;
+	private float price;
 	
-	public Possession(String name, Image image, Person owner, double price)
+	public Possession(String name, Image image, Person owner, float price)
 	{
 		super(name, image);
 		this.owner = owner;
 		this.price = price;
 	}
 	
-	public double getPrice()
+	public float getPrice()
 	{
 		return price; 
 	}
